Refactor: did a refactor of the finance controlling to occur at just one step below the cli entrance layer
